rename 'box description' config to 'indicator type'
