+++ python/config/indicator_type_matching.py
