+++ python/tests/aspace_data_indicator_type.json
