+++ python/tests/test_indicator_type_mapping.py
@@ -2,17 +2,15 @@ import unittest
 import os
 import json
 from config.base_match import match_containers
-from config.box_description_matching import (
-    get_alma_match_data as box_get_alma_match_data,
-    get_aspace_match_data as box_get_aspace_match_data,
-)
+from config.indicator_type_matching import get_alma_match_data, get_aspace_match_data
+
 
 # Get the directory of the test file
 current_dir = os.path.dirname(os.path.abspath(__file__))
 
 # Construct absolute paths for the test data files
-alma_data_path = os.path.join(current_dir, "alma_data_box.json")
-aspace_data_path = os.path.join(current_dir, "aspace_data_box.json")
+alma_data_path = os.path.join(current_dir, "alma_data_indicator_type.json")
+aspace_data_path = os.path.join(current_dir, "aspace_data_indicator_type.json")
 
 # Load the test data files
 with open(alma_data_path, "r") as alma_data_file:
@@ -22,15 +20,15 @@ with open(aspace_data_path, "r") as aspace_data_file:
     aspace_data = json.load(aspace_data_file)
 
 
-class TestBoxMapping(unittest.TestCase):
-    # use box_description_matching versions of get_data functions
+class TestIndicatorTypeMapping(unittest.TestCase):
+    # use indicator_type_description_matching versions of get_data functions
 
     def test_match_containers(self):
         # first item in alma_data should match first top container in aspace_data
         alma_items = [alma_data[0]["item_data"]]
         aspace_containers = [aspace_data[0]]
-        alma_match_data, items_with_duplicate_keys = box_get_alma_match_data(alma_items)
-        aspace_match_data, tcs_with_duplicate_keys = box_get_aspace_match_data(
+        alma_match_data, items_with_duplicate_keys = get_alma_match_data(alma_items)
+        aspace_match_data, tcs_with_duplicate_keys = get_aspace_match_data(
             aspace_containers
         )
         matched_aspace_containers, unhandled_data = match_containers(
@@ -52,8 +50,8 @@ class TestBoxMapping(unittest.TestCase):
         # second item in each set should not match
         alma_items = [alma_data[1]["item_data"]]
         aspace_containers = [aspace_data[1]]
-        alma_match_data, items_with_duplicate_keys = box_get_alma_match_data(alma_items)
-        aspace_match_data, tcs_with_duplicate_keys = box_get_aspace_match_data(
+        alma_match_data, items_with_duplicate_keys = get_alma_match_data(alma_items)
+        aspace_match_data, tcs_with_duplicate_keys = get_aspace_match_data(
             aspace_containers
         )
 
@@ -70,8 +68,8 @@ class TestBoxMapping(unittest.TestCase):
         # third item in each set should match, even though alma indicator has leading zeroes
         alma_items = [alma_data[2]["item_data"]]
         aspace_containers = [aspace_data[2]]
-        alma_match_data, items_with_duplicate_keys = box_get_alma_match_data(alma_items)
-        aspace_match_data, tcs_with_duplicate_keys = box_get_aspace_match_data(
+        alma_match_data, items_with_duplicate_keys = get_alma_match_data(alma_items)
+        aspace_match_data, tcs_with_duplicate_keys = get_aspace_match_data(
             aspace_containers
         )
         matched_aspace_containers, unhandled_data = match_containers(
@@ -94,8 +92,8 @@ class TestBoxMapping(unittest.TestCase):
         # even though alma indicator has " RESTRICTED" at the end
         alma_items = [alma_data[3]["item_data"]]
         aspace_containers = [aspace_data[3]]
-        alma_match_data, items_with_duplicate_keys = box_get_alma_match_data(alma_items)
-        aspace_match_data, tcs_with_duplicate_keys = box_get_aspace_match_data(
+        alma_match_data, items_with_duplicate_keys = get_alma_match_data(alma_items)
+        aspace_match_data, tcs_with_duplicate_keys = get_aspace_match_data(
             aspace_containers
         )
         matched_aspace_containers, unhandled_data = match_containers(
